Update ch3_challenge6.py
Hi!  Just started working with your book and I'm already glad I bought it!  

I was going through the challenge solutions and noticed that in your solution, you subtract 65 from the variable age , which will return a negative integer for anyone younger than 65.  Shouldn't it be the other way 'round?

Either way, it got me to join GitHub, so that's a plus!  

Thanks again for writing this book, I'm really enjoying it much better than the Codecademy tutorial.